Moved daw_json_assert to be a public'ly includable header.  This changes none of the interface but does add the possibility of including it. This helps when one is firewalling parse code into it's own TU
