+++ include/daw/json/impl/daw_json_parse_common.h
@@ -8,7 +8,7 @@
 
 #pragma once
 
-#include "daw_json_assert.h"
+#include "../daw_json_assert.h"
 #include "daw_json_exec_modes.h"
 #include "daw_json_parse_string_need_slow.h"
 #include "daw_json_parse_string_quote.h"
