+++ include/daw/json/impl/daw_json_parse_policy.h
@@ -8,7 +8,7 @@
 
 #pragma once
 
-#include "daw_json_assert.h"
+#include "../daw_json_assert.h"
 #include "daw_json_parse_common.h"
 #include "daw_json_parse_policy_cpp_comments.h"
 #include "daw_json_parse_policy_hash_comments.h"
