+++ include/daw/json/impl/daw_json_value.h
@@ -9,7 +9,7 @@
 #pragma once
 
 #include "daw_json_arrow_proxy.h"
-#include "daw_json_assert.h"
+#include "../daw_json_assert.h"
 #include "daw_json_parse_name.h"
 
 #include <ciso646>
