+++ tests/src/daw_json_link_test.cpp
@@ -10,7 +10,7 @@
 
 #include "daw/json/daw_json_iterator.h"
 #include "daw/json/daw_json_link.h"
-#include "daw/json/impl/daw_json_assert.h"
+#include "daw/json/daw_json_assert.h"
 #include "daw/json/impl/daw_json_exec_modes.h"
 
 #include <daw/daw_arith_traits.h>
